usArt_backend/requirements.txt
asgiref==3.5.2
django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.3.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
psycopg2-binary==2.9.4
pycodestyle==2.8.0
django-cors-headers==3.13.0
Pillow==9.3.0
mysqlclient==2.1.1